Cambio de vista de documentacion en admin
